Did some more tests. The preferred metric could be any of the following: - ROC_AUC - Recall - F1 - Balanced accuracy
balanced accuracy attributes 50% to the 0 class and 50% to the 1 class

using recall as scoring metric still yields only a recall of 33%, which is well above baseline (0% for all 0's) but that baseline would not be appropriate. 33% is still pretty low

with f1 we get ok scores, but here the precision metric is really high while the actual desired performance of the model is absent.